Renaming analyzers to decoders for consistency
