+++ src/multidecoder/decoders/network.py
