+++ src/multidecoder/decoders/path.py
