+++ src/multidecoder/decoders/pe_file.py
