+++ src/multidecoder/decoders/vba.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import regex as re
 
-from multidecoder.analyzers.concat import STRING_RE
+from multidecoder.decoders.concat import STRING_RE
 from multidecoder.hit import find_and_deobfuscate
 from multidecoder.node import Node
 from multidecoder.registry import decoder
