+++ tests/test_analyzers/test_pe_file.py
@@ -1,6 +1,6 @@
 import os
 
-from multidecoder.analyzers.pe_file import find_pe_files
+from multidecoder.decoders.pe_file import find_pe_files
 
 TEST_DIR = os.path.dirname(__file__)
