+++ tests/test_analyzers/test_xml.py
@@ -1,6 +1,6 @@
 import re
 
-from multidecoder.analyzers.xml import XML_ESCAPE_RE, find_xml_hex, unescape_xml
+from multidecoder.decoders.xml import XML_ESCAPE_RE, find_xml_hex, unescape_xml
 from multidecoder.node import Node
 
 external = b'TargetMode="&#x45;&#x78;&#x74;&#x65;&#x72;&#x6e;&#x61;&#x6c;"'
